[Refactor] 소수점으로 카운트하기 위한 수정
totalValue와 countValue를 정수에서 소수점으로 카운트.

+++ NumberCounter/NumberCounter/Model/NumberCounterModel.swift
@@ -8,8 +8,8 @@
 import Foundation
 
 struct NumberCounterModel {
-  var totalValue: Int
-  var countValue: Int
+  var totalValue: Double
+  var countValue: Double
   var changeValue: String
   var type: StateType
